[chat] refactor example folder
