--- applications/Chat/coati/ray/__init__.py
@@ -1,2 +1,0 @@
-from .src.detached_replay_buffer import DetachedReplayBuffer
-from .src.detached_trainer_ppo import DetachedPPOTrainer
