+++ applications/Chat/examples/ray/1m2t.sh
@@ -20,4 +20,4 @@ export RAY_NAMESPACE="admin"
 python 1m2t.py "/path/to/prompts.csv" --model gpt2 \
     --maker_strategy naive --trainer_strategy ddp --lora_rank 2 \
     --num_episodes 10 --max_timesteps 10 --update_timesteps 10 \
-    --max_epochs 10  #--debug 
+    --max_epochs 10  #--debug
