+++ applications/Chat/examples/ray/2m1t.sh
