+++ applications/Chat/examples/ray/2m2t.sh
